Dramatically improved the detection for the constructor injection refactoring (#3)
Previously this would only detect the refactoring under some special
circumstances. Now it works as expected for selecting either the type or
the property. It now also checks whether the method is the constructor.
Previously this would throw an exception if you tried to use the
refactor on a non-constructor method.